Add support for 64-bit event codes

+++ Source/Processors/SourceNode/SourceNode.h
@@ -110,7 +110,7 @@ private:
     DataBuffer* inputBuffer;
 
     uint64 timestamp;
-    int16* eventCodeBuffer;
+    uint64* eventCodeBuffer;
     int* eventChannelState;
